# src/main/java/com/example/bankcards/dto/ChangeCardStatusDto.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.enums.CardStatus;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class ChangeCardStatusDto {
    private CardStatus status;
    private Long cardId;
}